backend-api/requirements.txt
asgiref==3.5.0
astroid==2.11.2
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
colorama==0.4.4
dill==0.3.4
Django==4.0.3
django-cors-headers==3.12.0
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pylint==2.13.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.1
wrapt==1.14.0
